added User Profile

+++ app/Http/Controllers/PostController.php
@@ -30,10 +30,7 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        if($post->ownedBy(auth()->user())){
-            dd('no');
-        }
-        
+        $this->authorize('delete', $post);
         $post->delete();
 
         return back();
